+++ tests/test_pipeline_orchestrators.py
@@ -177,6 +177,8 @@ def test_pipeline_dlt():
         "edition": None,
         "libraries": None,
         "name": "pl-spark-dlt",
+        "name_prefix": None,
+        "name_suffix": None,
         "notifications": [],
         "photon": None,
         "serverless": None,